Conflicts:
	publiccms-parent/publiccms-common/src/main/java/com/publiccms/common/tools/ZipUtils.java
	publiccms-parent/publiccms-core/src/main/java/com/publiccms/logic/component/site/SiteComponent.java

+++ publiccms-parent/publiccms-core/src/main/java/com/publiccms/controller/admin/cms/CmsCategoryAdminController.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.publiccms.common.base.AbstractController;
+import com.publiccms.common.constants.CommonConstants;
 import com.publiccms.common.tools.CommonUtils;
 import com.publiccms.common.tools.ControllerUtils;
 import com.publiccms.common.tools.ExtendUtils;
@@ -39,8 +40,8 @@ import com.publiccms.logic.service.cms.CmsTagTypeService;
 import com.publiccms.logic.service.log.LogLoginService;
 import com.publiccms.logic.service.sys.SysExtendFieldService;
 import com.publiccms.logic.service.sys.SysExtendService;
-import com.publiccms.views.pojo.model.CmsCategoryModelParamters;
-import com.publiccms.views.pojo.model.CmsCategoryParamters;
+import com.publiccms.views.pojo.model.CmsCategoryModelParameters;
+import com.publiccms.views.pojo.model.CmsCategoryParameters;
 
 import freemarker.template.TemplateException;
 
@@ -77,20 +78,25 @@ public class CmsCategoryAdminController extends AbstractController {
     /**
      * @param entity
      * @param attribute
-     * @param categoryParamters
+     * @param categoryParameters
+     * @param _csrf
      * @param request
      * @param session
      * @param model
      * @return view name
      */
     @RequestMapping("save")
-    public String save(CmsCategory entity, CmsCategoryAttribute attribute, @ModelAttribute CmsCategoryParamters categoryParamters,
-            HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String save(CmsCategory entity, CmsCategoryAttribute attribute,
+            @ModelAttribute CmsCategoryParameters categoryParameters, String _csrf, HttpServletRequest request,
+            HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         SysSite site = getSite(request);
         if (null != entity.getId()) {
             CmsCategory oldEntity = service.getEntity(entity.getId());
             if (null == oldEntity || ControllerUtils.verifyNotEquals("siteId", site.getId(), oldEntity.getSiteId(), model)) {
-                return TEMPLATE_ERROR;
+                return CommonConstants.TEMPLATE_ERROR;
             }
             entity = service.update(entity.getId(), entity, ignoreProperties);
             if (null != entity) {
@@ -100,7 +106,7 @@ public class CmsCategoryAdminController extends AbstractController {
                 } else if (null != entity.getParentId() && null == oldEntity.getParentId()) {
                     service.generateChildIds(site.getId(), entity.getParentId());
                 }
-                logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+                logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                         LogLoginService.CHANNEL_WEB_MANAGER, "update.category", RequestUtils.getIpAddress(request),
                         CommonUtils.getDate(), JsonUtils.getString(entity)));
             }
@@ -111,37 +117,39 @@ public class CmsCategoryAdminController extends AbstractController {
             entity.setSiteId(site.getId());
             service.save(entity);
             service.addChildIds(entity.getParentId(), entity.getId());
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "save.category", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), JsonUtils.getString(entity)));
         }
-        if (null == extendService.getEntity(entity.getExtendId())) {
-            entity = service.updateExtendId(entity.getId(),
-                    (Integer) extendService.save(new SysExtend("category", entity.getId())));
-        }
 
-        Integer[] tagTypeIds = tagTypeService.update(site.getId(), categoryParamters.getTagTypes());
+        Integer[] tagTypeIds = tagTypeService.update(site.getId(), categoryParameters.getTagTypes());
         service.updateTagTypeIds(entity.getId(), arrayToCommaDelimitedString(tagTypeIds));// 更新保存标签分类
 
-        List<CmsCategoryModelParamters> categoryModelList = categoryParamters.getCategoryModelList();
+        List<CmsCategoryModelParameters> categoryModelList = categoryParameters.getCategoryModelList();
         if (CommonUtils.notEmpty(categoryModelList)) {
-            for (CmsCategoryModelParamters cmsCategoryModelParamters : categoryModelList) {
-                if (null != cmsCategoryModelParamters.getCategoryModel()) {
-                    cmsCategoryModelParamters.getCategoryModel().getId().setCategoryId(entity.getId());
-                    if (cmsCategoryModelParamters.isUse()) {
-                        categoryModelService.updateCategoryModel(cmsCategoryModelParamters.getCategoryModel());
+            for (CmsCategoryModelParameters cmsCategoryModelParameters : categoryModelList) {
+                if (null != cmsCategoryModelParameters.getCategoryModel()) {
+                    cmsCategoryModelParameters.getCategoryModel().getId().setCategoryId(entity.getId());
+                    if (cmsCategoryModelParameters.isUse()) {
+                        categoryModelService.updateCategoryModel(cmsCategoryModelParameters.getCategoryModel());
                     } else {
-                        categoryModelService.delete(cmsCategoryModelParamters.getCategoryModel().getId());
+                        categoryModelService.delete(cmsCategoryModelParameters.getCategoryModel().getId());
                     }
                 }
             }
         }
-        extendFieldService.update(entity.getExtendId(), categoryParamters.getContentExtends());// 修改或增加内容扩展字段
+        if (CommonUtils.notEmpty(categoryParameters.getContentExtends()) || CommonUtils.notEmpty(entity.getExtendId())) {
+            if (null == extendService.getEntity(entity.getExtendId())) {
+                entity = service.updateExtendId(entity.getId(),
+                        (Integer) extendService.save(new SysExtend("category", entity.getId())));
+            }
+            extendFieldService.update(entity.getExtendId(), categoryParameters.getContentExtends());// 修改或增加内容扩展字段
+        }
 
         CmsCategoryType categoryType = categoryTypeService.getEntity(entity.getTypeId());
         if (null != categoryType && CommonUtils.notEmpty(categoryType.getExtendId())) {
             List<SysExtendField> categoryTypeExtendList = extendFieldService.getList(categoryType.getExtendId());
-            Map<String, String> map = ExtendUtils.getSysExtentDataMap(categoryParamters.getExtendDataList(),
+            Map<String, String> map = ExtendUtils.getSysExtentDataMap(categoryParameters.getExtendDataList(),
                     categoryTypeExtendList);
             attribute.setData(ExtendUtils.getExtendString(map));
         } else {
@@ -154,31 +162,36 @@ public class CmsCategoryAdminController extends AbstractController {
             ControllerUtils.verifyCustom("static", true, model);
             log.error(e.getMessage(), e);
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
      * @param ids
      * @param parentId
+     * @param _csrf
      * @param request
      * @param session
      * @param model
      * @return view name
      */
     @RequestMapping("move")
-    public String move(Integer[] ids, Integer parentId, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String move(Integer[] ids, Integer parentId, String _csrf, HttpServletRequest request, HttpSession session,
+            ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         SysSite site = getSite(request);
         CmsCategory parent = service.getEntity(parentId);
         if (CommonUtils.notEmpty(ids) && (null == parent || null != parent && site.getId() == parent.getSiteId())) {
             for (Integer id : ids) {
                 move(site, id, parentId);
             }
-            logOperateService
-                    .save(new LogOperate(site.getId(), getAdminFromSession(session).getId(), LogLoginService.CHANNEL_WEB_MANAGER,
-                            "move.category", RequestUtils.getIpAddress(request), CommonUtils.getDate(),
-                            new StringBuilder(StringUtils.join(ids, ',')).append(" to ").append(parentId).toString()));
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
+                    LogLoginService.CHANNEL_WEB_MANAGER, "move.category", RequestUtils.getIpAddress(request),
+                    CommonUtils.getDate(),
+                    new StringBuilder(StringUtils.join(ids, ',')).append(" to ").append(parentId).toString()));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
@@ -200,13 +213,18 @@ public class CmsCategoryAdminController extends AbstractController {
     /**
      * @param ids
      * @param max
+     * @param _csrf
      * @param request
      * @param session
      * @param model
      * @return view name
      */
     @RequestMapping("publish")
-    public String publish(Integer[] ids, Integer max, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String publish(Integer[] ids, Integer max, String _csrf, HttpServletRequest request, HttpSession session,
+            ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         SysSite site = getSite(request);
         if (CommonUtils.notEmpty(ids)) {
             try {
@@ -216,34 +234,39 @@ public class CmsCategoryAdminController extends AbstractController {
             } catch (IOException | TemplateException e) {
                 ControllerUtils.verifyCustom("static", true, model);
                 log.error(e.getMessage(), e);
-                return TEMPLATE_ERROR;
+                return CommonConstants.TEMPLATE_ERROR;
             }
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "static.category", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), new StringBuilder(StringUtils.join(ids, ',')).append(",pageSize:")
                             .append((CommonUtils.empty(max) ? 1 : max)).toString()));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
      * @param id
      * @param typeId
+     * @param _csrf
      * @param request
      * @param session
      * @param model
      * @return view name
      */
     @RequestMapping("changeType")
-    public String changeType(Integer id, Integer typeId, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String changeType(Integer id, Integer typeId, String _csrf, HttpServletRequest request, HttpSession session,
+            ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         SysSite site = getSite(request);
         if (CommonUtils.notEmpty(id)) {
             service.changeType(id, typeId);
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "changeType.category", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), new StringBuilder(id).append(" to ").append(typeId).toString()));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
@@ -262,20 +285,25 @@ public class CmsCategoryAdminController extends AbstractController {
 
     /**
      * @param ids
+     * @param _csrf
      * @param request
      * @param session
+     * @param model
      * @return view name
      */
     @RequestMapping("delete")
-    public String delete(Integer[] ids, HttpServletRequest request, HttpSession session) {
+    public String delete(Integer[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         if (CommonUtils.notEmpty(ids)) {
             SysSite site = getSite(request);
             service.delete(site.getId(), ids);
             contentService.deleteByCategoryIds(site.getId(), ids);
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "delete.category", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), StringUtils.join(ids, ',')));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 }
