+++ publiccms-parent/publiccms-core/src/main/java/com/publiccms/controller/admin/cms/CmsDictionaryAdminController.java
@@ -6,10 +6,13 @@ import javax.servlet.http.HttpSession;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.publiccms.common.base.AbstractController;
+import com.publiccms.common.constants.CommonConstants;
 import com.publiccms.common.tools.CommonUtils;
+import com.publiccms.common.tools.ControllerUtils;
 import com.publiccms.common.tools.JsonUtils;
 import com.publiccms.common.tools.RequestUtils;
 import com.publiccms.entities.cms.CmsDictionary;
@@ -18,7 +21,7 @@ import com.publiccms.entities.sys.SysSite;
 import com.publiccms.logic.service.cms.CmsDictionaryDataService;
 import com.publiccms.logic.service.cms.CmsDictionaryService;
 import com.publiccms.logic.service.log.LogLoginService;
-import com.publiccms.views.pojo.model.CmsDictionaryParamters;
+import com.publiccms.views.pojo.model.CmsDictionaryParameters;
 
 /**
  *
@@ -29,51 +32,62 @@ import com.publiccms.views.pojo.model.CmsDictionaryParamters;
 @RequestMapping("cmsDictionary")
 public class CmsDictionaryAdminController extends AbstractController {
 
-    private String[] ignoreProperties = new String[] { "id" };
+    private String[] ignoreProperties = new String[] { "id", "siteId" };
 
     /**
      * @param entity
-     * @param dictionaryParamters
+     * @param dictionaryParameters
+     * @param _csrf
      * @param request
      * @param session
+     * @param model
      * @return view name
      */
     @RequestMapping("save")
-    public String save(CmsDictionary entity, CmsDictionaryParamters dictionaryParamters, HttpServletRequest request,
-            HttpSession session) {
+    public String save(CmsDictionary entity, CmsDictionaryParameters dictionaryParameters, String _csrf, HttpServletRequest request,
+            HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         SysSite site = getSite(request);
         if (null != entity.getId()) {
             entity = service.update(entity.getId(), entity, ignoreProperties);
-            dataService.update(entity.getId(), dictionaryParamters.getDataList());
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            dataService.update(entity.getId(), dictionaryParameters.getDataList());
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "update.cmsDictionary", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), JsonUtils.getString(entity)));
         } else {
+            entity.setSiteId(site.getId());
             service.save(entity);
-            dataService.save(entity.getId(), dictionaryParamters.getDataList());
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            dataService.save(entity.getId(), dictionaryParameters.getDataList());
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "save.cmsDictionary", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), JsonUtils.getString(entity)));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
      * @param ids
+     * @param _csrf
      * @param request
      * @param session
+     * @param model
      * @return view name
      */
     @RequestMapping("delete")
-    public String delete(Long[] ids, HttpServletRequest request, HttpSession session) {
+    public String delete(Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         SysSite site = getSite(request);
         if (CommonUtils.notEmpty(ids)) {
             service.delete(ids);
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "delete.cmsDictionary", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), StringUtils.join(ids, ',')));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     @Autowired
