+++ publiccms-parent/publiccms-core/src/main/java/com/publiccms/controller/admin/cms/CmsPlaceAdminController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.publiccms.common.base.AbstractController;
+import com.publiccms.common.constants.CommonConstants;
 import com.publiccms.common.tools.CommonUtils;
 import com.publiccms.common.tools.ControllerUtils;
 import com.publiccms.common.tools.ExtendUtils;
@@ -26,7 +27,7 @@ import com.publiccms.logic.component.template.TemplateComponent;
 import com.publiccms.logic.service.cms.CmsPlaceAttributeService;
 import com.publiccms.logic.service.cms.CmsPlaceService;
 import com.publiccms.logic.service.log.LogLoginService;
-import com.publiccms.views.pojo.model.CmsPlaceParamters;
+import com.publiccms.views.pojo.model.CmsPlaceParameters;
 
 /**
  * 
@@ -48,19 +49,26 @@ public class CmsPlaceAdminController extends AbstractController {
 
     /**
      * @param entity
-     * @param placeParamters
+     * @param placeParameters
+     * @param _csrf
      * @param request
      * @param session
      * @param model
      * @return view name
      */
     @RequestMapping(value = "save", method = RequestMethod.POST)
-    public String save(CmsPlace entity, @ModelAttribute CmsPlaceParamters placeParamters, HttpServletRequest request,
-            HttpSession session, ModelMap model) {
+    public String save(CmsPlace entity, @ModelAttribute CmsPlaceParameters placeParameters, String _csrf,
+            HttpServletRequest request, HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         if (null != entity && CommonUtils.notEmpty(entity.getPath())) {
-            entity.setPath(entity.getPath().replace("//", SEPARATOR));
+            if (!entity.getPath().startsWith(CommonConstants.SEPARATOR)) {
+                entity.setPath(CommonConstants.SEPARATOR + entity.getPath());
+            }
+            entity.setPath(entity.getPath().replace("//", CommonConstants.SEPARATOR));
             SysSite site = getSite(request);
-            Long userId = getAdminFromSession(session).getId();
+            Long userId = ControllerUtils.getAdminFromSession(session).getId();
             if (CommonUtils.empty(entity.getItemType()) || CommonUtils.empty(entity.getItemId())) {
                 entity.setItemType(CmsPlaceService.ITEM_TYPE_CUSTOM);
                 entity.setItemId(null);
@@ -68,7 +76,7 @@ public class CmsPlaceAdminController extends AbstractController {
             if (null != entity.getId()) {
                 CmsPlace oldEntity = service.getEntity(entity.getId());
                 if (null == oldEntity || ControllerUtils.verifyNotEquals("siteId", site.getId(), oldEntity.getSiteId(), model)) {
-                    return TEMPLATE_ERROR;
+                    return CommonConstants.TEMPLATE_ERROR;
                 }
                 entity = service.update(entity.getId(), entity, ignoreProperties);
                 if (null != entity) {
@@ -84,86 +92,103 @@ public class CmsPlaceAdminController extends AbstractController {
                         RequestUtils.getIpAddress(request), CommonUtils.getDate(), entity.getPath()));
             }
             String filePath = siteComponent.getWebTemplateFilePath(site, TemplateComponent.INCLUDE_DIRECTORY + entity.getPath());
-            Map<String, String> map = ExtendUtils.getExtentDataMap(placeParamters.getExtendDataList(),
+            Map<String, String> map = ExtendUtils.getExtentDataMap(placeParameters.getExtendDataList(),
                     metadataComponent.getPlaceMetadata(filePath).getExtendList());
             String extentString = ExtendUtils.getExtendString(map);
             attributeService.updateAttribute(entity.getId(), extentString);
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
      * @param ids
+     * @param _csrf
      * @param request
      * @param session
      * @param model
      * @return view name
      */
     @RequestMapping("refresh")
-    public String refresh(Long[] ids, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String refresh(Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         if (CommonUtils.notEmpty(ids)) {
             SysSite site = getSite(request);
             service.refresh(site.getId(), ids);
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "refresh.place", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), StringUtils.join(ids, ',')));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
      * @param ids
+     * @param _csrf
      * @param request
      * @param session
      * @param model
      * @return view name
      */
     @RequestMapping("check")
-    public String check(Long[] ids, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String check(Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         if (CommonUtils.notEmpty(ids)) {
             SysSite site = getSite(request);
             service.check(site.getId(), ids);
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "check.place", RequestUtils.getIpAddress(request), CommonUtils.getDate(),
                     StringUtils.join(ids, ',')));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
      * @param path
+     * @param _csrf
      * @param request
      * @param session
      * @param model
      * @return view name
      */
     @RequestMapping("clear")
-    public String clear(String path, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String clear(String path, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         if (CommonUtils.notEmpty(path)) {
             SysSite site = getSite(request);
             service.delete(site.getId(), path);
-            logOperateService
-                    .save(new LogOperate(site.getId(), getAdminFromSession(session).getId(), LogLoginService.CHANNEL_WEB_MANAGER,
-                            "clear.place", RequestUtils.getIpAddress(request), CommonUtils.getDate(), path));
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
+                    LogLoginService.CHANNEL_WEB_MANAGER, "clear.place", RequestUtils.getIpAddress(request), CommonUtils.getDate(),
+                    path));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 
     /**
      * @param ids
+     * @param _csrf
      * @param request
      * @param session
+     * @param model
      * @return view name
      */
     @RequestMapping("delete")
-    public String delete(Long[] ids, HttpServletRequest request, HttpSession session) {
+    public String delete(Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+        if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         if (CommonUtils.notEmpty(ids)) {
             SysSite site = getSite(request);
             service.delete(site.getId(), ids);
-            logOperateService.save(new LogOperate(site.getId(), getAdminFromSession(session).getId(),
+            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "delete.place", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), StringUtils.join(ids, ',')));
         }
-        return TEMPLATE_DONE;
+        return CommonConstants.TEMPLATE_DONE;
     }
 }
